refactor: reorganize project directories for improved separation of concerns
