+++ app/src/main/AndroidManifest.xml
@@ -13,7 +13,7 @@
         android:theme="@style/Theme.MetronomePlus"
         tools:targetApi="31">
         <activity
-            android:name=".ui.MainActivity"
+            android:name=".presenter.ui.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
