+++ app/src/main/java/br/com/jonatas/metronomeplus/domain/model/Beat.kt
@@ -1,4 +1,4 @@
-package br.com.jonatas.metronomeplus.model
+package br.com.jonatas.metronomeplus.domain.model
 
 data class Beat(val state: BeatState)
